OAK-7551 - Implement a stateless cancellation strategy
The cancellation strategy previously represented by `CancelCompactionSupplier`
has been replaced by `Canceller`. `Canceller` repesents a hierarchy of
cancellation conditions which can rely a cancellation request as soon as one of
the conditions is met. This allows the code to create more specific
cancellation conditions for specific parts of the codebase, withouth amassing
the cancellation state in a single entity.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1833822 13f79535-47bb-0310-9956-ffa450edef68